+++ bigquery-connector-common/src/main/java/com/google/cloud/bigquery/connector/common/BigQueryUtil.java
@@ -26,6 +26,7 @@ import com.google.cloud.bigquery.BigQueryException;
 import com.google.cloud.bigquery.Clustering;
 import com.google.cloud.bigquery.ExternalTableDefinition;
 import com.google.cloud.bigquery.Field;
+import com.google.cloud.bigquery.Field.Mode;
 import com.google.cloud.bigquery.FieldList;
 import com.google.cloud.bigquery.HivePartitioningOptions;
 import com.google.cloud.bigquery.LegacySQLTypeName;
@@ -301,8 +302,17 @@ public class BigQueryUtil {
     if (sourceField == destinationField) {
       return true;
     }
-    // if both are null we would have caught it earlier
-    if (sourceField == null || destinationField == null) {
+
+    // if the destination field is NULLABLE or REPEATED and there is no matching field in the source
+    // then it is
+    // supported
+    // but if the destination field is REQUIRED then we do need source field to write into it..
+    if (sourceField == null) {
+      return destinationField.getMode() != Mode.REQUIRED;
+    }
+
+    // cannot write if the destination table doesn't have the field
+    if (destinationField == null) {
       return false;
     }
 
@@ -316,7 +326,7 @@ public class BigQueryUtil {
     return Objects.equal(sourceField.getName(), destinationField.getName())
         && typeWriteable(sourceField.getType(), destinationField.getType())
         && (!enableModeCheckForSchemaFields
-            || Objects.equal(
+            || isModeWritable(
                 nullableIfNull(sourceField.getMode()), nullableIfNull(destinationField.getMode())))
         && ((sourceField.getMaxLength() == null && destinationField.getMaxLength() == null)
             || (sourceField.getMaxLength() != null
@@ -367,6 +377,18 @@ public class BigQueryUtil {
   }
 
   @VisibleForTesting
+  static boolean isModeWritable(Field.Mode sourceMode, Field.Mode destinationMode) {
+    switch (destinationMode) {
+      case REPEATED:
+        return sourceMode == Mode.REPEATED;
+      case REQUIRED:
+      case NULLABLE:
+        return sourceMode != Mode.REPEATED;
+    }
+    return false;
+  }
+
+  @VisibleForTesting
   static boolean fieldListWritable(
       FieldList sourceFieldList,
       FieldList destinationFieldList,
@@ -374,8 +396,14 @@ public class BigQueryUtil {
     if (sourceFieldList == destinationFieldList) {
       return true;
     }
+
     // if both are null we would have caught it earlier
     if (sourceFieldList == null || destinationFieldList == null) {
+      return false;
+    }
+
+    // cannot write of the source has more fields than the destination table.
+    if (sourceFieldList.size() > destinationFieldList.size()) {
       return false;
     }
 
@@ -385,9 +413,9 @@ public class BigQueryUtil {
         destinationFieldList.stream()
             .collect(Collectors.toMap(Field::getName, Function.identity()));
 
-    for (Map.Entry<String, Field> e : sourceFieldsMap.entrySet()) {
-      Field f1 = e.getValue();
-      Field f2 = destinationFieldsMap.get(e.getKey());
+    for (Map.Entry<String, Field> e : destinationFieldsMap.entrySet()) {
+      Field f1 = sourceFieldsMap.get(e.getKey());
+      Field f2 = e.getValue();
       if (!fieldWritable(f1, f2, enableModeCheckForSchemaFields)) {
         return false;
       }