+++ bigquery-connector-common/src/test/java/com/google/cloud/bigquery/connector/common/BigQueryUtilTest.java
@@ -263,9 +263,15 @@ public class BigQueryUtilTest {
         Schema.of(
             Field.newBuilder("foo", StandardSQLTypeName.STRING).setMode(Mode.REPEATED).build());
 
-    assertThat(BigQueryUtil.schemaWritable(s1, s2, false, true)).isFalse();
+    assertThat(BigQueryUtil.schemaWritable(s1, s1, false, true)).isTrue();
+    assertThat(BigQueryUtil.schemaWritable(s2, s2, false, true)).isTrue();
+    assertThat(BigQueryUtil.schemaWritable(s3, s3, false, true)).isTrue();
+    assertThat(BigQueryUtil.schemaWritable(s1, s2, false, true)).isTrue();
     assertThat(BigQueryUtil.schemaWritable(s1, s3, false, true)).isFalse();
     assertThat(BigQueryUtil.schemaWritable(s2, s3, false, true)).isFalse();
+    assertThat(BigQueryUtil.schemaWritable(s2, s1, false, true)).isTrue();
+    assertThat(BigQueryUtil.schemaWritable(s3, s1, false, true)).isFalse();
+    assertThat(BigQueryUtil.schemaWritable(s3, s2, false, true)).isFalse();
   }
 
   @Test
@@ -280,9 +286,48 @@ public class BigQueryUtilTest {
         Schema.of(
             Field.newBuilder("foo", StandardSQLTypeName.STRING).setMode(Mode.REPEATED).build());
 
+    assertThat(BigQueryUtil.schemaWritable(s1, s1, false, false)).isTrue();
+    assertThat(BigQueryUtil.schemaWritable(s2, s2, false, false)).isTrue();
+    assertThat(BigQueryUtil.schemaWritable(s3, s3, false, false)).isTrue();
     assertThat(BigQueryUtil.schemaWritable(s1, s2, false, false)).isTrue();
     assertThat(BigQueryUtil.schemaWritable(s1, s3, false, false)).isTrue();
     assertThat(BigQueryUtil.schemaWritable(s2, s3, false, false)).isTrue();
+    assertThat(BigQueryUtil.schemaWritable(s2, s1, false, false)).isTrue();
+    assertThat(BigQueryUtil.schemaWritable(s3, s1, false, false)).isTrue();
+    assertThat(BigQueryUtil.schemaWritable(s3, s2, false, false)).isTrue();
+  }
+
+  @Test
+  public void testSchemaWritableWithMoreUnEqualNumberOfFields() {
+    Schema s1 =
+        Schema.of(
+            Field.newBuilder("foo1", StandardSQLTypeName.STRING).setMode(Mode.NULLABLE).build(),
+            Field.newBuilder("foo2", StandardSQLTypeName.STRING).setMode(Mode.NULLABLE).build());
+    Schema s2 =
+        Schema.of(
+            Field.newBuilder("foo1", StandardSQLTypeName.STRING).setMode(Mode.NULLABLE).build(),
+            Field.newBuilder("foo2", StandardSQLTypeName.STRING).setMode(Mode.NULLABLE).build(),
+            Field.newBuilder("foo3", StandardSQLTypeName.STRING).setMode(Mode.NULLABLE).build());
+    Schema s3 =
+        Schema.of(
+            Field.newBuilder("foo1", StandardSQLTypeName.STRING).setMode(Mode.NULLABLE).build());
+
+    assertThat(BigQueryUtil.schemaWritable(s1, s2, false, true)).isTrue();
+    assertThat(BigQueryUtil.schemaWritable(s1, s3, false, true)).isFalse();
+    assertThat(BigQueryUtil.schemaWritable(s3, s2, false, true)).isTrue();
+  }
+
+  @Test
+  public void testIsModeWritable() {
+    assertThat(BigQueryUtil.isModeWritable(Mode.NULLABLE, Mode.NULLABLE)).isTrue();
+    assertThat(BigQueryUtil.isModeWritable(Mode.NULLABLE, Mode.REQUIRED)).isTrue();
+    assertThat(BigQueryUtil.isModeWritable(Mode.NULLABLE, Mode.REPEATED)).isFalse();
+    assertThat(BigQueryUtil.isModeWritable(Mode.REQUIRED, Mode.NULLABLE)).isTrue();
+    assertThat(BigQueryUtil.isModeWritable(Mode.REQUIRED, Mode.REQUIRED)).isTrue();
+    assertThat(BigQueryUtil.isModeWritable(Mode.REQUIRED, Mode.REPEATED)).isFalse();
+    assertThat(BigQueryUtil.isModeWritable(Mode.REPEATED, Mode.NULLABLE)).isFalse();
+    assertThat(BigQueryUtil.isModeWritable(Mode.REPEATED, Mode.REQUIRED)).isFalse();
+    assertThat(BigQueryUtil.isModeWritable(Mode.REPEATED, Mode.REPEATED)).isTrue();
   }
 
   @Test